fix: product form price, stock hint, and image size validation
- Price: allow any whole number (step=1, integer validation) so values like
  5132 pass; only decimals are rejected — no more 'nearest valid value' step
  errors forcing multiples of 100.
- Stock: show an inline amber hint when stock is 0 so users understand the
  product saves but won't be sellable until stock is added.
- Image: enforce the 2 MB limit client-side at selection, before the file is
  loaded into the cropper, with a clear message (server max:2048 unchanged).

+++ app/Http/Controllers/Tenant/ProductController.php
@@ -45,7 +45,7 @@ class ProductController extends Controller
             'category_id' => ['nullable', 'exists:categories,id'],
             'name'        => ['required', 'string', 'max:200'],
             'image'       => ['nullable', 'image', 'mimes:png,jpg,jpeg,webp', 'max:2048'],
-            'price'       => ['required', 'numeric', 'min:0'],
+            'price'       => ['required', 'integer', 'min:0'],
             'stock'       => ['required', 'integer', 'min:0'],
         ]);
 
@@ -85,7 +85,7 @@ class ProductController extends Controller
             'category_id' => ['nullable', 'exists:categories,id'],
             'name'        => ['required', 'string', 'max:200'],
             'image'       => ['nullable', 'image', 'mimes:png,jpg,jpeg,webp', 'max:2048'],
-            'price'       => ['required', 'numeric', 'min:0'],
+            'price'       => ['required', 'integer', 'min:0'],
             'stock'       => ['required', 'integer', 'min:0'],
         ]);
